rename generic HealthChecker package. Reversion poms to 0.3.0
