+++ legacy-caching-rest-client/pom.xml
@@ -5,12 +5,12 @@
     <parent>
         <artifactId>java-utils</artifactId>
         <groupId>fi.vm.sade.java-utils</groupId>
-        <version>0.2.2-SNAPSHOT</version>
+        <version>0.3.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>legacy-caching-rest-client</artifactId>
-    <version>0.2.0-SNAPSHOT</version>
+    <version>0.3.0-SNAPSHOT</version>
 
     <properties>
         <spring.version>5.1.5.RELEASE</spring.version>
@@ -100,7 +100,7 @@
         <dependency>
             <groupId>fi.vm.sade.java-utils</groupId>
             <artifactId>opintopolku-healthcheck-api</artifactId>
-            <version>0.1.0-SNAPSHOT</version>
+            <version>0.3.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>fi.vm.sade.java-utils</groupId>
