+++ opintopolku-healthcheck-api/pom.xml
@@ -4,11 +4,11 @@
     <parent>
         <groupId>fi.vm.sade.java-utils</groupId>
         <artifactId>java-utils</artifactId>
-        <version>0.2.2-SNAPSHOT</version>
+        <version>0.3.0-SNAPSHOT</version>
         <relativePath>..</relativePath>
     </parent>
 
     <artifactId>opintopolku-healthcheck-api</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
+    <version>0.3.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 </project>
