+++ opintopolku-spring-security/pom.xml
@@ -4,12 +4,12 @@
     <parent>
         <groupId>fi.vm.sade.java-utils</groupId>
         <artifactId>java-utils</artifactId>
-        <version>0.2.2-SNAPSHOT</version>
+        <version>0.3.0-SNAPSHOT</version>
         <relativePath>..</relativePath>
     </parent>
 
     <artifactId>opintopolku-spring-security</artifactId>
-    <version>0.2.0-SNAPSHOT</version>
+    <version>0.3.0-SNAPSHOT</version>
     <packaging>jar</packaging>
     <properties>
       <spring.security.version>4.2.11.RELEASE</spring.security.version>
