+++ spring-aware-health-check-servlet/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>fi.vm.sade.java-utils</groupId>
     <artifactId>java-utils</artifactId>
-    <version>0.2.2-SNAPSHOT</version>
+    <version>0.3.0-SNAPSHOT</version>
     <relativePath>..</relativePath>
   </parent>
   <artifactId>spring-aware-health-check-servlet</artifactId>
-  <version>0.2.0-SNAPSHOT</version>
+  <version>0.3.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <properties>
     <spring.version>5.1.5.RELEASE</spring.version>
@@ -75,12 +75,12 @@
       <dependency>
         <groupId>fi.vm.sade.java-utils</groupId>
         <artifactId>legacy-caching-rest-client</artifactId>
-        <version>0.2.0-SNAPSHOT</version>
+        <version>0.3.0-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>fi.vm.sade.java-utils</groupId>
         <artifactId>opintopolku-healthcheck-api</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
+        <version>0.3.0-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>fi.vm.sade.java-utils</groupId>
