+++ spring-aware-health-check-servlet/src/main/java/fi/vm/sade/javautils/healthcheck/ProxyAuthenticationChecker.java
@@ -1,6 +1,6 @@
 package fi.vm.sade.javautils.healthcheck;
 
-import fi.vm.sade.generic.healthcheck.HealthChecker;
+import fi.vm.sade.javautils.healthcheck.HealthChecker;
 import fi.vm.sade.javautils.legacy_caching_rest_client.CachingRestClient;
 import org.springframework.beans.factory.support.AbstractBeanFactory;
 import org.springframework.context.ApplicationContext;
